@@ -14,7 +14,7 @@
   },
   "homepage": "https://github.com/shuding/nextra-docs-template#readme",
   "dependencies": {
-    "next": "^13.0.6",
+    "next": "^14.2.14",
     "nextra": "latest",
     "nextra-theme-docs": "latest",
     "react": "^18.2.0",
