@@ -1,10 +1,8 @@
-const withNextra = require('nextra')({
-  theme: 'nextra-theme-docs',
-  themeConfig: './theme.config.tsx',
-})
+import nextra from 'nextra';
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  output: 'export',
   webpack: (
     config,
     { buildId, dev, isServer, defaultLoaders, nextRuntime, webpack }
@@ -12,16 +10,21 @@ const nextConfig = {
     if (config.cache && !dev) {
       config.cache = Object.freeze({
         type: 'memory',
-      })
+      });
     }
     // Important: return the modified config
-    return config
+    return config;
   },
-}
+};
 
-module.exports = {
+const withNextra = nextra({
+  theme: 'nextra-theme-docs',
+  themeConfig: './theme.config.tsx',
+});
+
+export default {
   ...withNextra(nextConfig),
   images: {
     unoptimized: true,
   },
-}
+};
