@@ -1,4 +1,4 @@
-{
+export default {
   "abstracting-entity-behavior": "Abstracting Entity Behavior",
-  "combat-done-right": "Combat Done Right",
-}
+  "combat-done-right": "Combat Done Right"
+};
